refactor: http package 구조 변경
