+++ src/main/java/controller/AbstractController.java
@@ -2,9 +2,9 @@ package controller;
 
 import java.util.Objects;
 
-import http.HttpRequest;
-import http.HttpResponse;
 import http.HttpStatus;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import servlet.HttpServlet;
 
 public abstract class AbstractController implements HttpServlet {
