+++ src/main/java/db/DataBase.java
@@ -1,7 +1,6 @@
 package db;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
