+++ src/main/java/http/request/HttpRequest.java
@@ -1,6 +1,10 @@
-package http;
+package http.request;
 
 import javax.annotation.Nullable;
+
+import http.HttpBody;
+import http.HttpHeaders;
+import http.HttpMethod;
 
 public interface HttpRequest {
     HttpMethod getMethod();
