+++ src/main/java/http/request/QueryParameters.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import java.util.Arrays;
 import java.util.Collections;
