+++ src/main/java/http/request/SimpleHttpRequest.java
@@ -1,9 +1,12 @@
-package http;
+package http.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.Objects;
 
+import http.HttpBody;
+import http.HttpHeaders;
+import http.HttpMethod;
 import utils.IOUtils;
 
 public class SimpleHttpRequest implements HttpRequest {
