+++ src/main/java/service/UserService.java
@@ -1,6 +1,5 @@
 package service;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
