+++ src/main/java/servlet/HttpServlet.java
@@ -1,7 +1,7 @@
 package servlet;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 
 public interface HttpServlet {
     void service(HttpRequest httpRequest, HttpResponse httpResponse);
