+++ src/main/java/webserver/RequestHandler.java
@@ -12,9 +12,9 @@ import java.nio.charset.StandardCharsets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import http.HttpRequest;
-import http.HttpResponse;
-import http.SimpleHttpRequest;
+import http.request.HttpRequest;
+import http.request.SimpleHttpRequest;
+import http.response.HttpResponse;
 import servlet.DispatcherServlet;
 import servlet.HttpServlet;
