+++ src/test/java/http/StartLineTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import http.request.StartLine;
+
 class StartLineTest {
 
     @Test
